fix gpt and wikipedia
